# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/signup": {
      "post": {
        "operationId": "AuthController_signup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignupResponse"
                }
              }
            }
          }
        },
        "summary": "Create an account. Terms must be accepted here.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "summary": "Sign in and receive an access + refresh token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponse"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new pair (rotating)",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "summary": "Revoke a single refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "operationId": "AuthController_logoutAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke every session for the current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Current user, including whether new terms need accepting",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/accept-terms": {
      "post": {
        "operationId": "AuthController_acceptTerms",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptTermsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept the current terms version",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangePasswordResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Change password and sign out every session",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/friends": {
      "get": {
        "operationId": "FriendshipsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "People you are connected to",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/requests": {
      "post": {
        "operationId": "FriendshipsController_send",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFriendRequestDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Send a friend request by email or user id",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/requests/incoming": {
      "get": {
        "operationId": "FriendshipsController_incoming",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Requests waiting for your answer",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/requests/outgoing": {
      "get": {
        "operationId": "FriendshipsController_outgoing",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FriendshipResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Requests you have sent",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/requests/{id}/accept": {
      "post": {
        "operationId": "FriendshipsController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accept a friend request",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/requests/{id}/reject": {
      "post": {
        "operationId": "FriendshipsController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reject a friend request",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/{userId}": {
      "delete": {
        "operationId": "FriendshipsController_remove",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Disconnect from someone (place access is unaffected)",
        "tags": [
          "friends"
        ]
      }
    },
    "/api/friends/{userId}/block": {
      "post": {
        "operationId": "FriendshipsController_block",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FriendshipResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Block someone from contacting you",
        "tags": [
          "friends"
        ]
      }
    },
...
  },
  "info": {
    "title": "FindMyStuff API",
    "description": "Remember where you kept your stuff.\n\nHierarchy: **Place** (home/office/locker) → **Storage** tree (room → almirah → shelf → box) → **Item**.\n\nEvery place-scoped route is nested under `/places/{placeId}` and is checked against your\nmembership of that place. Asking for a place you are not a member of returns **404**, not 403 —\nthe API never confirms that someone else's data exists.\n\nAll routes except signup, login, refresh, logout and health require\n`Authorization: Bearer <accessToken>`.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Home — Demo deleted"
          }
        },
        "required": [
          "success"
        ]
      },
      "SignupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "name": {
            "type": "string",
            "example": "Shristi Gupta"
          },
          "password": {
            "type": "string",
            "example": "Str0ng#Pass",
            "minLength": 8
          },
          "phone": {
            "type": "string",
            "example": "+919876543210"
          },
          "acceptTerms": {
            "type": "boolean",
            "description": "Must be true. The accepted version is recorded server-side.",
            "example": true
          }
        },
        "required": [
          "email",
          "name",
          "password",
          "acceptTerms"
        ]
      },
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "name": {
            "type": "string",
            "example": "Shristi Gupta"
          },
          "phone": {
            "type": "object",
            "nullable": true,
            "example": "+919876543210"
          },
          "avatarMediaId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "termsVersion": {
            "type": "object",
            "nullable": true,
            "example": "2026-09-01"
          },
          "termsAcceptedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "phone",
          "avatarMediaId",
          "termsVersion",
          "termsAcceptedAt",
          "createdAt"
        ]
      },
      "SignupResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Send as: Authorization: Bearer <accessToken>"
          },
          "refreshToken": {
            "type": "string",
            "description": "Single use — every refresh returns a new one"
          },
          "expiresIn": {
            "type": "string",
            "example": "15m",
            "description": "Lifetime of the access token"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "password": {
            "type": "string",
            "example": "Str0ng#Pass"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Send as: Authorization: Bearer <accessToken>"
          },
          "refreshToken": {
            "type": "string",
            "description": "Single use — every refresh returns a new one"
          },
          "expiresIn": {
            "type": "string",
            "example": "15m",
            "description": "Lifetime of the access token"
          },
          "user": {
            "$ref": "#/components/schemas/UserProfileResponse"
          },
          "termsAcceptanceRequired": {
            "type": "boolean",
            "example": false
          },
          "currentTermsVersion": {
            "type": "string",
            "example": "2026-09-01"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "user",
          "termsAcceptanceRequired",
          "currentTermsVersion"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token issued at login or last refresh"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "TokenPairResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "Send as: Authorization: Bearer <accessToken>"
          },
          "refreshToken": {
            "type": "string",
            "description": "Single use — every refresh returns a new one"
          },
          "expiresIn": {
            "type": "string",
            "example": "15m",
            "description": "Lifetime of the access token"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "name": {
            "type": "string",
            "example": "Shristi Gupta"
          },
          "phone": {
            "type": "object",
            "nullable": true,
            "example": "+919876543210"
          },
          "avatarMediaId": {
            "type": "object",
            "nullable": true,
            "format": "uuid"
          },
          "termsVersion": {
            "type": "object",
            "nullable": true,
            "example": "2026-09-01"
          },
          "termsAcceptedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "termsAcceptanceRequired": {
            "type": "boolean",
            "example": false,
            "description": "When true, show the terms dialog — other routes will return 403 until accepted"
          },
          "currentTermsVersion": {
            "type": "string",
            "example": "2026-09-01"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "phone",
          "avatarMediaId",
          "termsVersion",
          "termsAcceptedAt",
          "createdAt",
          "termsAcceptanceRequired",
          "currentTermsVersion"
        ]
      },
      "AcceptTermsDto": {
        "type": "object",
        "properties": {
          "accept": {
            "type": "boolean",
            "example": true
          },
          "version": {
            "type": "string",
            "description": "Version being accepted. Must match the current server version when provided."
          }
        },
        "required": [
          "accept"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string",
            "minLength": 8
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
      "ChangePasswordResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Password changed. Please sign in again."
          }
        },
        "required": [
          "success"
        ]
      },
...
      "UserCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "Shristi"
          },
          "email": {
            "type": "string",
            "example": "shristi@example.com"
          },
          "avatarMediaId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "FriendResponse": {
        "type": "object",
        "properties": {
          "friendshipId": {
            "type": "string",
            "format": "uuid"
          },
          "since": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "user": {
            "description": "The other person, whoever sent the request",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserCard"
              }
            ]
          }
        },
        "required": [
          "friendshipId",
          "since",
          "user"
        ]
      },
      "CreateFriendRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Either email or userId is required"
          },
          "userId": {
            "type": "string"
          },
          "message": {
            "type": "string",
            "example": "Hi Mummy, adding you so you can see the almirah"
          }
        }
      },
      "FriendshipResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED",
              "BLOCKED"
            ]
          },
          "message": {
            "type": "object",
            "nullable": true,
            "example": "Hi Mummy, adding you"
          },
          "requester": {
            "$ref": "#/components/schemas/UserCard"
          },
          "addressee": {
            "$ref": "#/components/schemas/UserCard"
          },
          "respondedAt": {
            "type": "object",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "message",
          "requester",
          "addressee",
          "respondedAt",
          "createdAt"
        ]
      },
...
    }
  }
}